Add player and gameboard objects.
This tic tac toe project requires that we make very modular code and avoid using too much global code. To do this, we have to make objects that represent a gameboard, a player(s), and the flow of the game. This commit adds the code for the gameboard and the players.