Add playGame function and game/player turn validation.
Keeping with the theme of modular code, this project requires that a playGame function be made to keep track of the flow of the game. Another requirement that became apparent along the way was validating if the game was won and if the players turn was valid. This commit will add a function to keep track of the game andvalidation for players/game condition. This makes the game fully functional from the console